Changed LONGs to floats. The define in mathieeesingtrans_intern.h does the trick.
git-svn-id: https://svn.aros.org/svn/aros/trunk@3962 fb15a70f-31f2-0310-bbcc-cdcc74a49acc

+++ workbench/libs/Attic/mathieeesptrans/ieeespacos.c
@@ -19,11 +19,11 @@
 
     NAME */
 
-      AROS_LH1(LONG, IEEESPAcos,
+      AROS_LH1(float, IEEESPAcos,
 
 /*  SYNOPSIS */
 
-      AROS_LHA(LONG, y, D0),
+      AROS_LHA(float, y, D0),
 
 /*  LOCATION */
