+++ workbench/libs/Attic/mathieeesptrans/ieeespasin.c
@@ -19,11 +19,11 @@
 
     NAME */
 
-      AROS_LH1(LONG, IEEESPAsin,
+      AROS_LH1(float, IEEESPAsin,
 
 /*  SYNOPSIS */
 
-      AROS_LHA(LONG, y, D0),
+      AROS_LHA(float, y, D0),
 
 /*  LOCATION */
