+++ workbench/libs/Attic/mathieeesptrans/ieeespsqrt.c
@@ -20,11 +20,11 @@
 
     NAME */
 
-      AROS_LH1(LONG, IEEESPSqrt,
+      AROS_LH1(float, IEEESPSqrt,
 
 /*  SYNOPSIS */
 
-      AROS_LHA(LONG, y, D0),
+      AROS_LHA(float, y, D0),
 
 /*  LOCATION */
