+++ workbench/libs/Attic/mathieeesptrans/ieeesptan.c
@@ -19,11 +19,11 @@
 
     NAME */
 
-      AROS_LH1(LONG, IEEESPTan,
+      AROS_LH1(float, IEEESPTan,
 
 /*  SYNOPSIS */
 
-      AROS_LHA(LONG, y, D0),
+      AROS_LHA(float, y, D0),
 
 /*  LOCATION */
