+++ workbench/libs/Attic/mathieeesptrans/ieeesptanh.c
@@ -20,11 +20,11 @@
 
     NAME */
 
-      AROS_LH1(LONG, IEEESPTanh,
+      AROS_LH1(float, IEEESPTanh,
 
 /*  SYNOPSIS */
 
-      AROS_LHA(LONG, y, D0),
+      AROS_LHA(float, y, D0),
 
 /*  LOCATION */
